menu-cli/pyproject.toml
[tool.poetry]
name = "menu"
version = "0.1.0"
description = ""
authors = ["Pete Jeffryes <pete.topleft@gmail.com>"]
readme = "README.md"
packages = [{include = "main.py"}]


[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.7.0"
boto = "^2.49.0"
requests = "^2.28.1"
python-slugify = "^7.0.0"
pydantic = "^1.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
menu = "main:app"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
black = "^22.10.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
mypy = "^0.991"
coverage = "^6.5.0"

# Tests
[tool.pytest.ini_options]
filterwarnings = "ignore::DeprecationWarning"

# Formatting and Linting
[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
ignore_missing_imports = "true"
disallow_untyped_defs = "true"
check_untyped_defs = "true"
warn_redundant_casts = "true"
no_implicit_optional = "true"
strict_optional = "true"

